Add new controller test files and refactor existing ones
Four new test files were added for FileUploadController, LoginController, ReservationController, and UserInfoController. Existing test files for CategoryController and PostController were refactored and moved to the ControllerTests folder for better organization. The main functionality and logic within the test files remain the same, with the sole intent of improving code structure.